Relax numpy version constraints
- Removes standalone numpy>=2 package entry
- Removes numpy<2 dependency constraints from jupyter,
  tensorflow, jax, jaxlib, flax, and numba packages
- Updates numba from 0.61.2 to 0.63.1

+++ dali/test/python/operator_1/test_inflate.py
@@ -79,7 +79,7 @@ def _test_sample_inflate(batch_size, np_dtype, seed):
 
             def sample(sample_size):
                 start = (sample_size - 1) * sample_size // 2
-                sample = np.arange(start, start + sample_size, dtype=np_dtype)
+                sample = np.arange(start, start + sample_size, dtype=np.int64).astype(np_dtype)
                 return sample, sample_to_lz4(sample)
 
             samples, deflated = list(zip(*[sample(sample_size) for sample_size in sample_sizes]))
